refactor: split battle value calculation modules
